refactor dataStream APIs into routers/logic.js, and libraries/dataStreamInterfaces.js
