+++ server.js
@@ -1873,38 +1873,6 @@ function objectWebServer() {
     webServer.get('/availableLogicBlocks/', function (req, res) {
         res.json(blockController.getLogicBlockList());
     });
-    
-    webServer.get('/availableDataSources/', function (req, res) {
-        let allAvailableDataSources = hardwareAPI.getAllAvailableDataSources();
-        res.json({
-            dataSources: allAvailableDataSources
-        });
-    });
-
-    webServer.get('/availableDataStreams/', function(req, res) {
-        let allAvailableDataStreams = hardwareAPI.getAllAvailableDataStreams();
-        res.json({
-            dataStreams: allAvailableDataStreams
-        });
-    });
-
-    webServer.post('/bindNodeToDataStream/', function(req, res) {
-        console.log('bindNodeToDataStream', req.body);
-        hardwareAPI.bindNodeToDataStream(req.body);
-        res.status(200).json({ success: true, error: null });
-    });
-    
-    webServer.post('/addDataSourceToInterface/', function(req, res) {
-        console.log('addDataSourceToInterface', req.body);
-        hardwareAPI.addDataSourceToInterface(req.body.interfaceName, req.body.dataSource);
-        res.status(200).json({ success: true, error: null });
-    });
-    
-    webServer.delete('/deleteDataSourceFromInterface/', function(req, res) {
-        console.log('deleteDataSourceFromInterface', req.body);
-        hardwareAPI.deleteDataSourceFromInterface(req.body.interfaceName, req.body.dataSource);
-        res.status(200).json({ success: true, error: null });
-    });
 
     // TODO: is the developer flag ever not true anymore? is it still useful to have?
     if (globalVariables.developer === true) {
